+++ src/commands/build.rs
@@ -176,7 +176,7 @@ impl Subcommand<Context> for Build {
         // plan for the provided target
         Plan::run(
             &current_project,
-            target_dir,
+            Some(target_dir),
             target,
             catalog,
             self.force,
@@ -187,6 +187,7 @@ impl Subcommand<Context> for Build {
             false,
             is_all,
             auto_discover,
+            false,
             envs,
             c.are_units_private_by_default(),
         )?;